signaling late is working, but have to test different stops late signaling and autoreset

+++ module/handlers_basic_commands.py
@@ -2,7 +2,7 @@ import logging
 from telegram import Update
 from telegram.ext import ContextTypes,CommandHandler,CallbackQueryHandler, MessageHandler
 from module.retrieve_webdata import getdownload_urls
-from module.markups import general_markup,bus_markup,tr_markup,transport_markup,times_markup,near_time_markup,any_markup
+from module.markups import general_markup,bus_markup,tr_markup,transport_markup,times_markup,near_time_markup,any_markup,backmarkup
 from module.timetables_operations.calculate_times import find_lines
 from module.timetables_operations.db import insert_db_user,select_db_users, insert_ritardo
 from module.timetables_operations.extract_excel import all_replacing
@@ -129,15 +129,10 @@ async def buttons(message: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
         else:
             await reset(context,message)
     
-    elif(context.chat_data['segnalaritardo']==1):
-        context.chat_data['segnalaritardo']=0
-        with sql.connect("fce_lines.db") as connection:
-            await insert_ritardo(connection,query.data,context,message)
-    
     elif(query.data == "ritardo"):
         await context.bot.delete_message(chat_id=message.effective_chat.id,message_id=message.effective_message.id) 
         context.chat_data['segnalaritardo']=1
-        await context.bot.send_message(chat_id=message.effective_chat.id,text="Manda l'ID della tratta che vuoi segnalare")
+        await context.bot.send_message(chat_id=message.effective_chat.id,text="Manda l'ID della tratta che vuoi segnalare",reply_markup=await backmarkup(),disable_web_page_preview=True)
 
     
     
@@ -180,8 +175,16 @@ Esempio di range: se scelgo le 8.00 intendiamo come range fino alle 8.59
 Si informano inoltre gli utenti che usando il bot si acconsente a memorizzare il chat_id riferito a questa chat e l'ora dell'ultimo utilizzo, il tutto a scopi puramente statistici e/o per ottenere degli avvisi a seguito di modifiche o altre informazioni (la memorizzazione é temporanea, se non si utilizza il bot per più di 2 mesi quei dati verranno cancellati automaticamente).""")
 
 async def scraping_messages(message: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    if message.message.text not in ["/start","/help","/contributors","/chat_id"] and context.chat_data['segnalaritardo']!=1:
+    if message.message.text not in ["/start","/help","/contributors","/chat_id"] and 'segnalaritardo' in context.chat_data and context.chat_data['segnalaritardo']!=1:
         await context.bot.delete_message(chat_id=message.effective_chat.id,message_id=message.effective_message.id)
+    
+    elif(context.chat_data and context.chat_data['segnalaritardo']==1):
+        context.chat_data['segnalaritardo']=0
+        with sql.connect("fce_lines.db") as connection:
+            await insert_ritardo(connection,message.message.text,context,message)
+            context.chat_data['counter']=6
+        
+
     if message.message.text == "/start":
         
         with sql.connect("users.db") as connection:
